requirements.txt
pyodbc
qrcode[pil]
Pillow
stripe


#pip install -r requirements.txt
